pyproject.toml
[project]
name = "code-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bert-score>=0.3.13",
    "boto3>=1.42.33",
    "datasets>=4.4.2",
    "litellm>=1.80.16",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
]
